LPS-90120 Use a MockMessage so code is not repeated.
