+++ modules/apps/portal-vulcan/portal-vulcan-test/src/testIntegration/java/com/liferay/portal/vulcan/internal/context/provider/test/AcceptLanguageContextProviderTest.java
@@ -33,25 +33,16 @@ import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.vulcan.context.AcceptLanguage;
+import com.liferay.portal.vulcan.internal.context.provider.test.util.MockMessage;
 
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.Locale;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
-
-import javax.servlet.http.HttpServletRequest;
 
 import javax.ws.rs.core.HttpHeaders;
 
-import org.apache.cxf.interceptor.InterceptorChain;
 import org.apache.cxf.jaxrs.ext.ContextProvider;
-import org.apache.cxf.message.Attachment;
-import org.apache.cxf.message.Exchange;
-import org.apache.cxf.message.Message;
-import org.apache.cxf.transport.Destination;
 
 import org.junit.AfterClass;
 import org.junit.Assert;
@@ -117,161 +108,6 @@ public class AcceptLanguageContextProviderTest {
 		_testCreateContext(LocaleUtil.TAIWAN, user);
 	}
 
-	private Message _createMessage(HttpServletRequest httpServletRequest) {
-		return new Message() {
-
-			@Override
-			public void clear() {
-			}
-
-			@Override
-			public boolean containsKey(Object key) {
-				return false;
-			}
-
-			@Override
-			public boolean containsValue(Object value) {
-				return false;
-			}
-
-			@Override
-			public Set<Entry<String, Object>> entrySet() {
-				return null;
-			}
-
-			@Override
-			public <T> T get(Class<T> clazz) {
-				return null;
-			}
-
-			@Override
-			public Object get(Object key) {
-				return null;
-			}
-
-			@Override
-			public Collection<Attachment> getAttachments() {
-				return null;
-			}
-
-			@Override
-			public <T> T getContent(Class<T> clazz) {
-				return null;
-			}
-
-			@Override
-			public Set<Class<?>> getContentFormats() {
-				return null;
-			}
-
-			@Override
-			public Object getContextualProperty(String contextProperty) {
-				if (Objects.equals(contextProperty, "HTTP.REQUEST")) {
-					return httpServletRequest;
-				}
-
-				return null;
-			}
-
-			@Override
-			public Set<String> getContextualPropertyKeys() {
-				return null;
-			}
-
-			@Override
-			public Destination getDestination() {
-				return null;
-			}
-
-			@Override
-			public Exchange getExchange() {
-				return null;
-			}
-
-			@Override
-			public String getId() {
-				return null;
-			}
-
-			@Override
-			public InterceptorChain getInterceptorChain() {
-				return null;
-			}
-
-			@Override
-			public boolean isEmpty() {
-				return false;
-			}
-
-			@Override
-			public Set<String> keySet() {
-				return null;
-			}
-
-			@Override
-			public <T> void put(Class<T> clazz, T t) {
-			}
-
-			@Override
-			public Object put(String key, Object value) {
-				return null;
-			}
-
-			@Override
-			public void putAll(Map<? extends String, ?> map) {
-			}
-
-			@Override
-			public <T> T remove(Class<T> clazz) {
-				return null;
-			}
-
-			@Override
-			public Object remove(Object key) {
-				return null;
-			}
-
-			@Override
-			public <T> void removeContent(Class<T> clazz) {
-			}
-
-			@Override
-			public void resetContextCache() {
-			}
-
-			@Override
-			public void setAttachments(Collection<Attachment> collection) {
-			}
-
-			@Override
-			public <T> void setContent(Class<T> clazz, Object object) {
-			}
-
-			@Override
-			public void setExchange(Exchange exchange) {
-			}
-
-			@Override
-			public void setId(String id) {
-			}
-
-			@Override
-			public void setInterceptorChain(InterceptorChain interceptorChain) {
-			}
-
-			@Override
-			public int size() {
-				return 0;
-			}
-
-			@Override
-			public Collection<Object> values() {
-				return null;
-			}
-
-		};
-	}
-
 	private void _testCreateContext(Locale userLocale, User user)
 		throws Exception {
 
@@ -279,7 +115,7 @@ public class AcceptLanguageContextProviderTest {
 
 		AcceptLanguage acceptLanguage =
 			_acceptLanguageContextProvider.createContext(
-				_createMessage(
+				new MockMessage(
 					new AcceptLanguageMockHttpServletRequest(
 						user, Locale.JAPAN)));
 
@@ -288,7 +124,7 @@ public class AcceptLanguageContextProviderTest {
 		// Three locales
 
 		acceptLanguage = _acceptLanguageContextProvider.createContext(
-			_createMessage(
+			new MockMessage(
 				new AcceptLanguageMockHttpServletRequest(
 					user, Locale.GERMAN, Locale.JAPAN, Locale.US)));
 
@@ -299,7 +135,7 @@ public class AcceptLanguageContextProviderTest {
 		Assert.assertEquals(userLocale, user.getLocale());
 
 		acceptLanguage = _acceptLanguageContextProvider.createContext(
-			_createMessage(new AcceptLanguageMockHttpServletRequest(user)));
+			new MockMessage(new AcceptLanguageMockHttpServletRequest(user)));
 
 		Assert.assertEquals(
 			user.getLocale(), acceptLanguage.getPreferredLocale());
